requirements.txt
asgiref==3.5.0
django-modeltranslation==0.18.2
django-nested-admin==3.4.0
django-parler==2.3
Pillow==9.1.0
python-monkey-business==1.0.0
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
